\name{dataFAQ}
\alias{dataFAQ}
\alias{dataFAQlog1}
\alias{dataFAQscale1}
\alias{dataFAQscale2}
\docType{data}

\title{Datasets for the FAQ}

\description{
Datasets used in the FAQ vignette.
}

\usage{
data(dataFAQlog1)
data(dataFAQscale1)
data(dataFAQscale2)
}

\format{ 
\code{dataFAQlog1}
\code{dataFAQscale1}
\code{dataFAQscale2} are vectors of numeric data.
}

\author{ 
Marie-Laure Delignette-Muller and Christophe Dutang.
}



\keyword{ datasets }% at least one, from doc/KEYWORDS
